+++ BasiliskII/src/cdrom.cpp
@@ -627,8 +627,8 @@ int16 CDROMPrime(uint32 pb, uint32 dce)
 	}
 	
 	// Update ParamBlock and DCE
-	WriteMacInt32(pb + ioActCount, actual);
-	WriteMacInt32(dce + dCtlPosition, ReadMacInt32(dce + dCtlPosition) + actual);
+	WriteMacInt32(pb + ioActCount, (uint32)actual);
+	WriteMacInt32(dce + dCtlPosition, ReadMacInt32(dce + dCtlPosition) + (uint32)actual);
 	return noErr;
 }
